# lib/api-zod/src/generated/types/clip.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceSegment } from './sourceSegment';

export interface Clip {
  id: string;
  index: number;
  title: string;
  hook: string;
  reason: string;
  duration: number;
  downloadUrl: string;
  previewUrl: string;
  sourceSegments?: SourceSegment[];
  subtitleUrl?: string;
  styledSubtitleUrl?: string;
}