# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadResponseMetadata } from './uploadResponseMetadata';

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadResponseMetadata;
}